heuristicSortPredicateSet completed
It has been added a heuristic that orders in some intelligent way the
predicate sets that can appear in the Coffee Server example. In
particular, it orders the ‘Served’ predicates of the Goal State by
creating simultaneously two arraylists, one with a sequence of the
nearest ’Served’ offices starting  at the initial point, the other  an
analogous sequence starting at the final point.